services/worker/requirements.txt
arq>=0.26
redis>=5.0
sqlalchemy>=2.0
asyncpg>=0.29
httpx>=0.26
feedparser>=6.0
google-generativeai>=0.4
beautifulsoup4>=4.12
lxml>=5.0
sentry-sdk>=1.40
